M-sized-integers step 3: `int` is deleted, and every model that writes it gets the repair — docs/debrief/QUEUE.md
The author's ruling, 2026-08-12: "i64 sempre e mai int, togli int". Not an
alias — the word is gone. And the reason is §1.1's rather than taste: `int`
carries forty years of conflicting widths, so a reader must know the platform
to know what it means, while `i64` is ambiguous to nobody. The 545 sites this
cost to migrate were 545 places carrying a number nobody could read off the
name.

565 occurrences of the whole word across 125 .hero files, the spec, design.md,
99 Rust files, and `to_int` -> `to_i64` in 21 more. Records under docs/ left
untouched: CLAUDE.md §14 does not rewrite a record.

`int` joins the foreign-word registry, which is where it belongs — it is the
word in that table a model is likeliest to write, and the repair is a pure
word-for-word swap, so it is `certain` and `heroes check --in-place` migrates a
whole file:

    error[reserved_word]: `int` is not a word in this language — an integer
    says its width: `i64`, and also `i8` `i16` `i32` `u8` `u16` `u32` `u64`
      fix (certain): replace `int` with `i64`

ONLY `int`. A first draft also reserved `integer`, `long`, `short`, `byte`,
`uint`, `usize` and `isize`, and that was a guess about the world rather than a
fact about the value — CLAUDE.md §11's named failure mode, committed by me in
the same session that found three instances of it in the tree. `byte` and
`short` are plausible identifiers (a lexer writes `byte = s[i]`) and `integer`
is an ordinary English word this project's own specification uses in prose.
Principle 0 admits a form when something needs it; nothing needed those seven.

Two things deliberately NOT renamed. `TokenKind::Int` still prints `int` in a
token dump: it names a TOKEN, not a type, and changing it churns every lexer
snapshot for zero semantics. And `int_out_of_range` keeps its code — "an
integer literal out of range" is true at all eight widths, so the name survives
the milestone it would otherwise have to be renamed for.

The typehash moved, and the instrument built for it caught it. mangle::instance
hashes the RENDERED type arguments, so `map<int, str>` and `map<i64, str>` are
different names in the generated C — every mangled name in every program using
a generic changed. tests/golden/run/premise-mangler-hash-in-heroes.hero
computes the same four values in Heroes and disagreed with the Rust side within
seconds of the rename. Both sides updated, and verified three ways: the Rust
mangler, the Heroes program, and an independent recomputation of the
polynomial. GOLDEN DIFF, read and quoted (§9):

    -1726907847  -75036336  -353437510  -1143528872
    +452710609   +947353345 +1680613593 +323221308

One of the four inputs is not a rename: `pair<int_str, x>` became
`pair<i64_str, x>` by hand. The word-boundary rename correctly skipped it —
`int` there is followed by `_` — but that string is half of the pair proving a
READABLE suffix would collide where a hash does not (`pair<i64, str_x>` and
`pair<i64_str, x>` both flatten to `pair_i64_str_x`). Left as `int_str` the
collision it demonstrates would no longer be reachable.

Two self-inflicted defects, both found and fixed here, both worth recording
because they are the shape of a large mechanical edit. A blanket `"int"` ->
`"i64"` rewrote the registry's own KEY, so the compiler rejected `i64` saying it
is not a word in this language. And a blanket `: int` -> `: i64` corrupted two
comments into `panic: i64eger overflow` — grepped for by the signature of the
damage rather than by re-reading 99 files.

Spec 2745 -> 2768, +23, and the delta is entirely the tokeniser: cl100k spends
one token on ` int` and two on `i64`, and the document names an integer 21
times. No rule added, none removed. Registered prediction is panel 042's #4.

472 tests green, clippy clean, examples/sqlite and examples/calculator both
run, `heroes mutate` unchanged.

+++ crates/heroes/src/resolve/tests/mutation.rs
@@ -8,13 +8,13 @@ fn a_cell_and_a_mutable_parameter_may_be_written() {
     assert_clean(
         "\
 record Lex
-    pos: int
+    pos: i64
 
 function advance(@l: Lex)
     l.pos @ l.pos + 1
 
-function count(n: int) -> int
-    total: int @ 0
+function count(n: i64) -> i64
+    total: i64 @ 0
     while total < n
         total @ total + 1
     return total
@@ -46,7 +46,7 @@ fn a_parameter_without_the_marker_cannot_be_written() {
     assert_eq!(
         diagnostics(
             "\
-function bump(n: int) -> int
+function bump(n: i64) -> i64
     n @ n + 1
     return n
 "
@@ -60,7 +60,7 @@ fn a_loop_variable_is_bound_afresh_and_is_not_a_cell() {
     assert_eq!(
         diagnostics(
             "\
-function main(xs: [int])
+function main(xs: [i64])
     for x in xs
         x @ x + 1
 "
@@ -74,7 +74,7 @@ fn there_are_no_mutable_globals() {
     assert_eq!(
         diagnostics(
             "\
-constant MAX: int
+constant MAX: i64
     64
 
 function main()
@@ -91,8 +91,8 @@ function main()
 fn the_index_of_an_element_write_is_a_read() {
     assert_clean(
         "\
-function main(n: int)
-    xs: [int] @ [0, 0]
+function main(n: i64)
+    xs: [i64] @ [0, 0]
     i = n
     xs[i] @ 1
     print(xs)
@@ -107,8 +107,8 @@ fn a_typo_on_the_left_of_a_mutation_is_an_unknown_name() {
     assert_eq!(
         diagnostics(
             "\
-function sum(xs: [int]) -> int
-    total: int @ 0
+function sum(xs: [i64]) -> i64
+    total: i64 @ 0
     for x in xs
         totl @ total + x
     return total
